Removed unnecessary catches.
Moved CommandWriterThread to its own file.
